Created files for classes. Created code for User.java
Created skeletal code for most of the classes.

Created code for User.java. User now has 3 elements: username, hashed_password and password_salt. There is method to create a new salt during construction, store the salt and hash the password with the salt and store it. Also created a method to hash an input string and compare with the hashed password.